Apply review suggestion
